Code refactor and small changes
